1A/requirements.txt
PyMuPDF
pdfminer.six==20221105
